refactor: separate auth and oauth
