+++ src/main/java/co/orange/ddanzi/controller/AuthController.java
@@ -4,7 +4,8 @@ import co.orange.ddanzi.domain.user.enums.LoginType;
 import co.orange.ddanzi.dto.auth.SigninRequestDto;
 import co.orange.ddanzi.dto.auth.VerifyRequestDto;
 import co.orange.ddanzi.global.common.response.ApiResponse;
-import co.orange.ddanzi.service.AuthService;
+import co.orange.ddanzi.service.auth.AuthService;
+import co.orange.ddanzi.service.auth.OAuthService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/v1/auth")
 public class AuthController {
     private final AuthService authService;
+    private final OAuthService oAuthService;
 
     @PostMapping("/signin/test")
     ApiResponse<?> test(@RequestBody SigninRequestDto requestDto){
@@ -26,12 +28,13 @@ public class AuthController {
     @PostMapping("/signin")
     ApiResponse<?> signin(@RequestBody SigninRequestDto requestDto) throws JsonProcessingException {
         if(requestDto.getType().equals(LoginType.KAKAO))
-            return authService.kakaoSignIn(requestDto.getToken());
-        return authService.kakaoSignIn(requestDto.getToken());
+            return oAuthService.kakaoSignIn(requestDto.getToken());
+        return oAuthService.kakaoSignIn(requestDto.getToken());
     }
 
     @PostMapping("/verification")
     ApiResponse<?> verify(@RequestBody VerifyRequestDto requestDto) {
         return authService.verify(requestDto);
     }
+
 }
